+++ litho-widget/src/main/java/com/facebook/litho/widget/TextSpec.java
@@ -47,6 +47,7 @@ import android.text.TextUtils.TruncateAt;
 import android.text.style.ClickableSpan;
 import android.text.style.ImageSpan;
 import android.view.View;
+import androidx.annotation.Dimension;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
 import androidx.core.text.TextDirectionHeuristicCompat;
@@ -166,6 +167,8 @@ import com.facebook.yoga.YogaDirection;
 class TextSpec {
 
   private static final Typeface DEFAULT_TYPEFACE = Typeface.DEFAULT;
+  private static final @Dimension(unit = Dimension.SP) int DEFAULT_TEXT_SIZE_SP = 14;
+  private static final int UNSET = -1;
   private static final int DEFAULT_COLOR = 0;
   private static final String TAG = "TextSpec";
   private static final String WRONG_TEXT_SIZE = "TextSpec:WrongTextSize";
@@ -188,7 +191,7 @@ class TextSpec {
       new ColorStateList(
           DEFAULT_TEXT_COLOR_STATE_LIST_STATES, DEFAULT_TEXT_COLOR_STATE_LIST_COLORS);
 
-  @PropDefault protected static final int textSize = 13;
+  @PropDefault protected static final int textSize = UNSET;
   @PropDefault protected static final int textStyle = DEFAULT_TYPEFACE.getStyle();
   @PropDefault protected static final Typeface typeface = DEFAULT_TYPEFACE;
   @PropDefault protected static final float spacingMultiplier = 1.0f;
@@ -323,6 +326,7 @@ class TextSpec {
 
     Layout newLayout =
         createTextLayout(
+            context,
             widthSpec,
             ellipsize,
             shouldIncludeFontPadding,
@@ -405,6 +409,7 @@ class TextSpec {
   }
 
   private static Layout createTextLayout(
+      ComponentContext context,
       int widthSpec,
       TruncateAt ellipsize,
       boolean shouldIncludeFontPadding,
@@ -474,7 +479,6 @@ class TextSpec {
         .setShadowLayer(shadowRadius, shadowDx, shadowDy, shadowColor)
         .setSingleLine(isSingleLine)
         .setText(text)
-        .setTextSize(textSize)
         .setWidth(SizeSpec.getSize(widthSpec), textMeasureMode)
         .setIncludeFontPadding(shouldIncludeFontPadding)
         .setTextSpacingExtra(extraSpacing)
@@ -483,6 +487,14 @@ class TextSpec {
         .setJustificationMode(justificationMode)
         .setBreakStrategy(breakStrategy)
         .setHyphenationFrequency(hyphenationFrequency);
+
+    // text size must be set before the line hight
+    if (textSize != UNSET) {
+      layoutBuilder.setTextSize(textSize);
+    } else {
+      int defaultTextSize = context.getResourceResolver().sipsToPixels(DEFAULT_TEXT_SIZE_SP);
+      layoutBuilder.setTextSize(defaultTextSize);
+    }
 
     if (lineHeight != Float.MAX_VALUE) {
       layoutBuilder.setLineHeight(lineHeight);
@@ -631,6 +643,7 @@ class TextSpec {
     } else {
       textLayout.set(
           createTextLayout(
+              c,
               SizeSpec.makeSizeSpec((int) layoutWidth, EXACTLY),
               ellipsize,
               shouldIncludeFontPadding,
@@ -691,6 +704,7 @@ class TextSpec {
 
         Layout newLayout =
             createTextLayout(
+                c,
                 SizeSpec.makeSizeSpec((int) layoutWidth, EXACTLY),
                 ellipsize,
                 shouldIncludeFontPadding,